third_party/WebKit/LayoutTests/html/dialog/inert-focus-in-frames-expected.txt

Tests inert node focusing across frames and iframes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
Opening a modal dialog in frame1. It blocks other nodes in its document.
PASS "frame1-input"; focusedElement === theElement is false
PASS "iframe-input"; focusedElement === theElement is false
Even a modal dialog in the iframe is blocked by the modal dialog in the parent frame1.
PASS "iframe-dialog-button"; focusedElement === theElement is false
An iframe within a modal dialog can still be focused.
PASS "iframe-under-dialog-input"; focusedElement === theElement is true
A modal dialog does not block nodes in a sibling frame.
PASS "frame2-div"; focusedElement === theElement is true
Closing the dialog in frame1. The modal dialog in the iframe does not block nodes in its parent.
PASS "iframe-input"; focusedElement === theElement is false
PASS "frame1-input"; focusedElement === theElement is true
